requirements.txt
feedparser>=6.0.11
certifi>=2024.2.2
python-dotenv>=1.0.1
anthropic>=0.40.0
